E2E, test_odf_query: extract
