+++ src/e2e/app/cli/inmem/tests/tests/rest_api/mod.rs
@@ -11,6 +11,7 @@ mod test_accounts;
 mod test_auth;
 mod test_dataset;
 mod test_odf_core;
+mod test_odf_query;
 mod test_openapi;
 mod test_swagger;
 mod test_upload;
