+++ src/e2e/app/cli/inmem/tests/tests/rest_api/test_dataset.rs
@@ -13,49 +13,6 @@ use kamu_cli_e2e_common::prelude::*;
 
 kamu_cli_run_api_server_e2e_test!(
     storage = inmem,
-    fixture = kamu_cli_e2e_repo_tests::rest_api::test_dataset_tail_st,
-    // We need synthetic time for the tests, but the third-party JWT code
-    // uses the current time. Assuming that the token lifetime is 24 hours, we will
-    // use the projected date (the current day) as a workaround.
-    options = Options::default().with_today_as_frozen_system_time(),
-    extra_test_groups = "engine, datafusion"
-);
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-kamu_cli_run_api_server_e2e_test!(
-    storage = inmem,
-    fixture = kamu_cli_e2e_repo_tests::rest_api::test_dataset_tail_mt,
-    // We need synthetic time for the tests, but the third-party JWT code
-    // uses the current time. Assuming that the token lifetime is 24 hours, we will
-    // use the projected date (the current day) as a workaround.
-    options = Options::default()
-        .with_multi_tenant()
-        .with_today_as_frozen_system_time(),
-    extra_test_groups = "engine, datafusion"
-);
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-kamu_cli_run_api_server_e2e_test!(
-    storage = inmem,
-    fixture = kamu_cli_e2e_repo_tests::rest_api::test_dataset_metadata_st,
-);
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-kamu_cli_run_api_server_e2e_test!(
-    storage = inmem,
-    fixture = kamu_cli_e2e_repo_tests::rest_api::test_dataset_metadata_mt,
-    options = Options::default()
-        .with_multi_tenant()
-        .with_today_as_frozen_system_time(),
-);
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-kamu_cli_run_api_server_e2e_test!(
-    storage = inmem,
     fixture = kamu_cli_e2e_repo_tests::rest_api::test_datasets_by_id,
 );
